Adding a parse API service
